Move sources + resources into TroposCore
- Marks appropriate types and members as public.
- Moves localized strings into the TroposCore framework bundle.
