+++ Tropos/Formatters/PrecipitationChanceFormatter.swift
@@ -1,4 +1,5 @@
 import Foundation
+import TroposCore
 
 struct PrecipitationChanceFormatter {
     func localizedStringFromPrecipitation(precipitation: Precipitation) -> String {
