+++ TroposCore/Sources/Extensions/NSBundle.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension NSBundle {
+    static var troposBundle: NSBundle {
+        return NSBundle(forClass: TroposBundleClass.self)
+    }
+}
+
+private class TroposBundleClass: NSObject {}
