+++ TroposCore/Sources/Extensions/NSLocalizedString.swift
@@ -1,0 +1,5 @@
+import func Foundation.NSLocalizedString
+
+func TroposCoreLocalizedString(string: String, comment: String = "") -> String {
+    return NSLocalizedString(string, bundle: .troposBundle, comment: comment)
+}
