+++ TroposCore/Specs/Models/CardinalDirectionSpec.swift
@@ -1,4 +1,4 @@
-@testable import Tropos
+import TroposCore
 import Quick
 import Nimble
