+++ TroposCore/Specs/Resources/New York.placemark
