--- TroposCore/Specs/TroposCoreSpecs.swift
@@ -1,11 +1,0 @@
-@testable import TroposCore
-import Quick
-import Nimble
-
-final class TroposCoreSpecs: QuickSpec {
-    override func spec() {
-        it("works") {
-            expect(1 + 1) == 2
-        }
-    }
-}
